requirements.txt
asgiref==3.3.1
Django==3.1.4
django-appconf==1.0.4
djangorestframework==3.12.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.4
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
